add websockets
if you update a story or column, it will write a messsage
to the websocket

all websocket connections are stored in a map with the board_name
as key:

{ 'gz' => [Connection1, Connection2], 'uboot' => [Connection1] }

Connection is a scruct with the websocket und a BoardStruct.

BoardStruct is a nested struct with the columns of the board.
The Columns are nested structs with the stories of the column.