Archive Feature
Use Archive Attribute of Story to remove a Story form Board
without permanently deleting it.
For now we only filter the stories, and there is no way
to retrive them. In a later change I will add a list of
Archived Stories. We still need to know what a user want to do
with this.

closes #206